a low memory footprint Gate document adapter (to list of sentences)
The purpose of this implementation is not to have an efficient
computation profile, but a low memory footprint; so the extraction
algorithm can be (and is) low-performing.

Further reduction on memory footprint could be pursued:
- (though unlikely) not all tokens actually belong to a sentence; these
tokens could be skipped, since they will never be reachable;
- from item above, it follows that the sentences will be directly next
to one another; there will be no need to store references to start AND
end offsets: either will be enough.