LowMemoryFootprintDocument to adapt LightDocument instead of Gate doc
As a side effect, unneeded tokens will not be stored in feature matrix
AND sentencesStart was rendered unused (since any given sentence always
closely follows the previous one, with exception of the first, which, by
virtue of being the first, starts on the first token).

On the coding side, this separation conveys responsibility more
elegantly. LightDocument knows to read Gate document and computes the
necessary offsets. LowMemoryFootprintDocument, on the other side, can
adapt any List of Sentences, LightDocument being just one of them.