Token officially became Map<String, Object>
That is a mapping of features (described as string, valued as any type
of object).

The most dangerous change of semantics lies on Token.put(). Previously
it would return "this" to allow for method chaining (fluent interface),
but Map.put() mandates another behavior. This modification led to
changes mainly on test cases.

Among this patch's most silly side-effects is a mind-boggling
interaction bug between mockito and guava's ForwardingMap, which was
left partially unresolved on CompositeTagger.FilteringToken (traces of
this sh*t can also be found on CompositeTaggerBehavior).

Minor changes:
* Token lost its getFeatures() since it was rendered unnecessary;
* dependency on guava, for its ForwardingMap proved to be very useful
  (and thus got pervasively used);
* a new implementation of Token: MapWrapToken, which happens to be
  simply a bridge between bare Maps and Token (note that Token is just a
  tagger interface, so this "implementation" really does nothing but to
  tag a Map as Token);
* NULL_TOKEN itself changed to wrap an emptyMap();
* DefaultToken is just a HashMap.